seeder

+++ database/migrations/2023_03_22_170554_add_service_provider_id_to_prices_table.php
@@ -14,7 +14,7 @@ class AddServiceProviderIdToPricesTable extends Migration
     public function up()
     {
         Schema::table('prices', function (Blueprint $table) {
-            $table->integer('service_provider_id')->nullable();
+            $table->bigInteger('service_provider_id')->nullable();
         });
     }
