+++ database/migrations/2023_03_22_172202_add_service_provider_id_to_settings_table.php
@@ -14,7 +14,7 @@ class AddServiceProviderIdToSettingsTable extends Migration
     public function up()
     {
         Schema::table('settings', function (Blueprint $table) {
-            $table->integer('service_provider_id')->nullable();
+            $table->bigInteger('service_provider_id')->nullable();
         });
     }
